+++ crates/env/src/engine/on_chain/pallet_revive.rs
@@ -880,6 +880,15 @@ impl TypedEnvBackend for EnvInstance {
         scale::Decode::decode(&mut &output[64..96]).expect("must exist")
     }
 
+    #[cfg(feature = "unstable-hostfn")]
+    fn to_account_id<E: Environment>(&mut self, addr: Address) -> E::AccountId {
+        let mut scope = self.scoped_buffer();
+        let account_id: &mut [u8; 32] = scope.take(32).try_into().unwrap();
+        ext::to_account_id(addr.as_fixed_bytes(), account_id);
+        scale::Decode::decode(&mut &account_id[..])
+            .expect("A contract being executed must have a valid account id.")
+    }
+
     fn address(&mut self) -> Address {
         let mut scope = self.scoped_buffer();