+++ integration-tests/public/runtime-call-contract/sandbox-runtime/pallet-revive-caller/src/executor.rs
@@ -5,18 +5,18 @@ use frame_support::{
 };
 use frame_system::pallet_prelude::OriginFor;
 use ink::{
-    abi::AbiEncodeWith,
-    env::{
-        call::{
-            utils::DecodeMessageResult,
-            ExecutionInput,
-            Executor,
-        },
-        Environment,
-    },
-    primitives::U256,
     Address,
     MessageResult,
+    abi::AbiEncodeWith,
+    env::{
+        Environment,
+        call::{
+            ExecutionInput,
+            Executor,
+            utils::DecodeMessageResult,
+        },
+    },
+    primitives::U256,
 };
 use pallet_revive::{
     DepositLimit,
@@ -25,8 +25,6 @@ use pallet_revive::{
 use sp_runtime::traits::Bounded;
 
 pub struct PalletReviveExecutor<E: Environment, Runtime: pallet_revive::Config> {
-    // todo
-    //pub origin: AccountIdOf<Runtime>,
     pub origin: OriginFor<Runtime>,
     pub contract: Address,
     pub value: BalanceOf<Runtime>,
@@ -41,7 +39,6 @@ impl<E, R> Executor<E> for PalletReviveExecutor<E, R>
 where
     E: Environment,
     R: pallet_revive::Config,
-
     BalanceOf<R>: Into<U256> + TryFrom<U256> + Bounded,
     MomentOf<R>: Into<U256>,
     <R as frame_system::Config>::Hash: IsType<sp_runtime::testing::H256>,